\name{CRN.DAILY.URL}
\alias{CRN.DAILY.URL}
\docType{data}
\title{ URL of the daily data collection
 
}
\description{URL to the parent directory
 
}
\usage{ CRN.DAILY.URL }
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/daily01/"
}
\details{ This is the base url. Subdirectories below this are read
   and files in those directories are downloaded
 
}
\source{
 \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/daily01/}
}
\references{
 \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/daily01/README.txt}
}
\examples{
print(CRN.DAILY.URL)
 
}
\keyword{urls}
