\name{CRN.HOURLY.URL}
\alias{CRN.HOURLY.URL}
\docType{data}
\title{URL of the hourly data collection
 
}
\description{URL to the parent directory
 
}
\usage{ CRN.HOURLY.URL }
\format{
  The format is:
 chr "ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/hourly02/"
}
\details{This is the base url. Subdirectories below this are read
   and files in those directories are downloaded
 
}
\source{
 \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/hourly02/}
}
\references{
 \url{ftp://ftp.ncdc.noaa.gov/pub/data/uscrn/products/hourly02/README.txt}
}
\examples{
print(CRN.HOURLY.URL)
 
}
\keyword{urls}
