\name{DAILY_DIR}
\alias{DAILY_DIR}
\docType{data}
\title{ Directory for downloaded files of daily data
 
}
\description{The daily files are downloaded from the various subdirectories
 on the ftp and they all are downloaded to the DAILY directory. These files
 are then collated into a monolithic file
 
}
\usage{ DAILY_DIR }
\format{
  The format is:
 chr "DailySourceData"
}
 
\examples{
 print(DAILY_DIR)
 file.exists(DAILY_DIR)
}
\keyword{directories}
