\name{HOURLY_DIR}
\alias{HOURLY_DIR}
\docType{data}
\title{ Directory for Hourly data
 
}
\description{The hourly files are downloaded from the various subdirectories
 on the ftp and they all are downloaded to the HOURLY directory. These files
 are then collated into a monolithic file using \code{collateHourly}
 
}
\usage{ HOURLY_DIR }
\format{
  The format is:
 chr "HourlySourceData"
}
 
 
\examples{
print(HOURLY_DIR)
 
}
\keyword{directories}
